refactor: correct the imports for DataEntry
