+++ src/core_modules/capture-core/components/Pages/ViewEvent/ViewEventComponent/ViewEvent.container.js
@@ -5,7 +5,7 @@ import {
 } from './viewEvent.actions';
 import { ViewEventComponent } from './ViewEvent.component';
 import { getDataEntryKey } from '../../../DataEntry/common/getDataEntryKey';
-import { editEventIds } from '../../EditEvent/DataEntry/editEventDataEntry.actions';
+import { editEventIds } from '../../../WidgetEventEdit/DataEntry/editEventDataEntry.actions';
 import { viewEventIds } from '../EventDetailsSection/eventDetails.actions';
 import { withErrorMessageHandler } from '../../../../HOC/withErrorMessageHandler';
 import { makeProgramStageSelector, makeEventAccessSelector } from './viewEvent.selectors';
