+++ src/core_modules/capture-core/components/WidgetEventEdit/DataEntry/fieldValidators/eventDate.validatorContainersGetter.js
@@ -1,7 +1,7 @@
 // @flow
 import { hasValue } from 'capture-core-utils/validators/form';
 import i18n from '@dhis2/d2-i18n';
-import { isValidDate } from '../../../../../utils/validators/form';
+import { isValidDate } from '../../../../utils/validators/form';
 
 const preValidateDate = (value?: ?string) => {
     if (!value) {
