+++ src/core_modules/capture-core/components/WidgetEventEdit/EditEventDataEntry/EditEventDataEntry.component.js
@@ -2,9 +2,9 @@
 import React, { Component } from 'react';
 import { withStyles } from '@material-ui/core/styles';
 import i18n from '@dhis2/d2-i18n';
-import { getEventDateValidatorContainers } from '../../Pages/EditEvent/DataEntry/fieldValidators/eventDate.validatorContainersGetter';
+import { getEventDateValidatorContainers } from '../DataEntry/fieldValidators/eventDate.validatorContainersGetter';
 import type { RenderFoundation } from '../../../metaData';
-import { withMainButton } from '../../Pages/EditEvent/DataEntry/withMainButton';
+import { withMainButton } from '../DataEntry/withMainButton';
 import { withFilterProps } from '../../FormFields/New/HOC/withFilterProps';
 
 import {
@@ -32,7 +32,7 @@ import {
 } from '../../FormFields/New';
 
 import { inMemoryFileStore } from '../../DataEntry/file/inMemoryFileStore';
-import labelTypeClasses from '../../Pages/EditEvent/DataEntry/dataEntryFieldLabels.module.css';
+import labelTypeClasses from '../DataEntry/dataEntryFieldLabels.module.css';
 
 const getStyles = (theme: Theme) => ({
     dataEntryContainer: {
