+++ src/core_modules/capture-core/reducers/descriptions/editEvent.reducerDescription.js
@@ -1,7 +1,7 @@
 // @flow
 import { createReducerDescription } from '../../trackerRedux/trackerReducer';
 import { actionTypes as editEventActionTypes } from '../../components/Pages/ViewEvent/ViewEventComponent/editEvent.actions';
-import { actionTypes as editEventDataEntryActionTypes } from '../../components/Pages/EditEvent/DataEntry/editEventDataEntry.actions';
+import { actionTypes as editEventDataEntryActionTypes } from '../../components/WidgetEventEdit/DataEntry/editEventDataEntry.actions';
 
 export const editEventPageDesc = createReducerDescription({
     [editEventActionTypes.EDIT_EVENT_FROM_URL]: (state, action) => {
